small changes are related to renaming bug_034
